code optimized and added publications to  mega menu
